telegram/bot-test/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
anyio==4.6.2.post1
    # via httpx
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
h11==0.14.0
    # via httpcore
httpcore==1.0.6
    # via httpx
httpx==0.27.2
    # via python-telegram-bot
idna==3.10
    # via
    #   anyio
    #   httpx
python-dotenv==1.0.1
    # via -r requirements.in
python-telegram-bot==21.7
    # via -r requirements.in
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
